TUSK_URLS are considered (#87)
* TUSK_URLS are considered

* Update rEADME
